+++ src/Symfony/Component/Cache/Adapter/PhpArrayAdapter.php
@@ -13,10 +13,12 @@ namespace Symfony\Component\Cache\Adapter;
 
 use Psr\Cache\CacheItemInterface;
 use Psr\Cache\CacheItemPoolInterface;
+use Symfony\Component\Cache\CacheInterface;
 use Symfony\Component\Cache\CacheItem;
 use Symfony\Component\Cache\Exception\InvalidArgumentException;
 use Symfony\Component\Cache\PruneableInterface;
 use Symfony\Component\Cache\ResettableInterface;
+use Symfony\Component\Cache\Traits\GetTrait;
 use Symfony\Component\Cache\Traits\PhpArrayTrait;
 
 /**
@@ -26,9 +28,10 @@ use Symfony\Component\Cache\Traits\PhpArrayTrait;
  * @author Titouan Galopin <galopintitouan@gmail.com>
  * @author Nicolas Grekas <p@tchwork.com>
  */
-class PhpArrayAdapter implements AdapterInterface, PruneableInterface, ResettableInterface
+class PhpArrayAdapter implements AdapterInterface, CacheInterface, PruneableInterface, ResettableInterface
 {
     use PhpArrayTrait;
+    use GetTrait;
 
     private $createCacheItem;
 
@@ -36,11 +39,10 @@ class PhpArrayAdapter implements AdapterInterface, PruneableInterface, Resettabl
      * @param string           $file         The PHP file were values are cached
      * @param AdapterInterface $fallbackPool A pool to fallback on when an item is not hit
      */
-    public function __construct($file, AdapterInterface $fallbackPool)
+    public function __construct(string $file, AdapterInterface $fallbackPool)
     {
         $this->file = $file;
         $this->pool = $fallbackPool;
-        $this->zendDetectUnicode = ini_get('zend.detect_unicode');
         $this->createCacheItem = \Closure::bind(
             function ($key, $value, $isHit) {
                 $item = new CacheItem();
@@ -56,19 +58,17 @@ class PhpArrayAdapter implements AdapterInterface, PruneableInterface, Resettabl
     }
 
     /**
-     * This adapter should only be used on PHP 7.0+ to take advantage of how PHP
-     * stores arrays in its latest versions. This factory method decorates the given
-     * fallback pool with this adapter only if the current PHP version is supported.
+     * This adapter takes advantage of how PHP stores arrays in its latest versions.
      *
      * @param string                 $file         The PHP file were values are cached
-     * @param CacheItemPoolInterface $fallbackPool Fallback for old PHP versions or opcache disabled
+     * @param CacheItemPoolInterface $fallbackPool Fallback when opcache is disabled
      *
      * @return CacheItemPoolInterface
      */
     public static function create($file, CacheItemPoolInterface $fallbackPool)
     {
-        // Shared memory is available in PHP 7.0+ with OPCache enabled and in HHVM
-        if ((\PHP_VERSION_ID >= 70000 && ini_get('opcache.enable')) || defined('HHVM_VERSION')) {
+        // Shared memory is available in PHP 7.0+ with OPCache enabled
+        if (ini_get('opcache.enable')) {
             if (!$fallbackPool instanceof AdapterInterface) {
                 $fallbackPool = new ProxyAdapter($fallbackPool);
             }
@@ -77,6 +77,31 @@ class PhpArrayAdapter implements AdapterInterface, PruneableInterface, Resettabl
         }
 
         return $fallbackPool;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function get(string $key, callable $callback, float $beta = null)
+    {
+        if (null === $this->values) {
+            $this->initialize();
+        }
+        if (null === $value = $this->values[$key] ?? null) {
+            if ($this->pool instanceof CacheInterface) {
+                return $this->pool->get($key, $callback, $beta);
+            }
+
+            return $this->doGet($this->pool, $key, $callback, $beta ?? 1.0);
+        }
+        if ('N;' === $value) {
+            return null;
+        }
+        if (\is_string($value) && isset($value[2]) && ':' === $value[1]) {
+            return unserialize($value);
+        }
+
+        return $value;
     }
 
     /**
@@ -101,12 +126,8 @@ class PhpArrayAdapter implements AdapterInterface, PruneableInterface, Resettabl
             $value = null;
         } elseif (\is_string($value) && isset($value[2]) && ':' === $value[1]) {
             try {
-                $e = null;
                 $value = unserialize($value);
-            } catch (\Error $e) {
-            } catch (\Exception $e) {
-            }
-            if (null !== $e) {
+            } catch (\Throwable $e) {
                 $value = null;
                 $isHit = false;
             }
@@ -226,10 +247,7 @@ class PhpArrayAdapter implements AdapterInterface, PruneableInterface, Resettabl
         return $this->pool->commit();
     }
 
-    /**
-     * @return \Generator
-     */
-    private function generateItems(array $keys)
+    private function generateItems(array $keys): \Generator
     {
         $f = $this->createCacheItem;
         $fallbackKeys = array();
@@ -243,9 +261,7 @@ class PhpArrayAdapter implements AdapterInterface, PruneableInterface, Resettabl
                 } elseif (\is_string($value) && isset($value[2]) && ':' === $value[1]) {
                     try {
                         yield $key => $f($key, unserialize($value), true);
-                    } catch (\Error $e) {
-                        yield $key => $f($key, null, false);
-                    } catch (\Exception $e) {
+                    } catch (\Throwable $e) {
                         yield $key => $f($key, null, false);
                     }
                 } else {