+++ src/Symfony/Component/Cache/Simple/PhpFilesCache.php
@@ -20,22 +20,15 @@ class PhpFilesCache extends AbstractCache implements PruneableInterface
     use PhpFilesTrait;
 
     /**
-     * @param string      $namespace
-     * @param int         $defaultLifetime
-     * @param string|null $directory
-     *
      * @throws CacheException if OPcache is not enabled
      */
-    public function __construct($namespace = '', $defaultLifetime = 0, $directory = null)
+    public function __construct(string $namespace = '', int $defaultLifetime = 0, string $directory = null)
     {
-        if (!static::isSupported()) {
-            throw new CacheException('OPcache is not enabled');
-        }
+        self::$startTime = self::$startTime ?? $_SERVER['REQUEST_TIME'] ?? time();
         parent::__construct('', $defaultLifetime);
         $this->init($namespace, $directory);
 
         $e = new \Exception();
         $this->includeHandler = function () use ($e) { throw $e; };
-        $this->zendDetectUnicode = ini_get('zend.detect_unicode');
     }
 }