+++ src/Symfony/Component/DependencyInjection/Compiler/ResolveBindingsPass.php
@@ -59,7 +59,7 @@ class ResolveBindingsPass extends AbstractRecursivePass
      */
     protected function processValue($value, $isRoot = false)
     {
-        if ($value instanceof TypedReference && $value->getType() === $this->container->normalizeId($value)) {
+        if ($value instanceof TypedReference && $value->getType() === (string) $value) {
             // Already checked
             $bindings = $this->container->getDefinition($this->currentId)->getBindings();
 
@@ -83,7 +83,7 @@ class ResolveBindingsPass extends AbstractRecursivePass
                 $this->unusedBindings[$bindingId] = array($key, $this->currentId);
             }
 
-            if (isset($key[0]) && '$' === $key[0]) {
+            if (preg_match('/^(?:(?:array|bool|float|int|string) )?\$/', $key)) {
                 continue;
             }
 
@@ -123,15 +123,21 @@ class ResolveBindingsPass extends AbstractRecursivePass
                     continue;
                 }
 
+                $typeHint = ProxyHelper::getTypeHint($reflectionMethod, $parameter);
+
+                if (array_key_exists($k = ltrim($typeHint, '\\').' $'.$parameter->name, $bindings)) {
+                    $arguments[$key] = $this->getBindingValue($bindings[$k]);
+
+                    continue;
+                }
+
                 if (array_key_exists('$'.$parameter->name, $bindings)) {
                     $arguments[$key] = $this->getBindingValue($bindings['$'.$parameter->name]);
 
                     continue;
                 }
 
-                $typeHint = ProxyHelper::getTypeHint($reflectionMethod, $parameter, true);
-
-                if (!isset($bindings[$typeHint])) {
+                if (!$typeHint || '\\' !== $typeHint[0] || !isset($bindings[$typeHint = substr($typeHint, 1)])) {
                     continue;
                 }