+++ src/Symfony/Component/DependencyInjection/EnvVarProcessor.php
@@ -36,10 +36,12 @@ class EnvVarProcessor implements EnvVarProcessorInterface
             'base64' => 'string',
             'bool' => 'bool',
             'const' => 'bool|int|float|string|array',
+            'csv' => 'array',
             'file' => 'string',
             'float' => 'float',
             'int' => 'int',
             'json' => 'array',
+            'key' => 'bool|int|float|string|array',
             'resolve' => 'string',
             'string' => 'string',
         );
@@ -51,6 +53,25 @@ class EnvVarProcessor implements EnvVarProcessorInterface
     public function getEnv($prefix, $name, \Closure $getEnv)
     {
         $i = strpos($name, ':');
+
+        if ('key' === $prefix) {
+            if (false === $i) {
+                throw new RuntimeException(sprintf('Invalid configuration: env var "key:%s" does not contain a key specifier.', $name));
+            }
+
+            $next = substr($name, $i + 1);
+            $key = substr($name, 0, $i);
+            $array = $getEnv($next);
+
+            if (!is_array($array)) {
+                throw new RuntimeException(sprintf('Resolved value of "%s" did not result in an array value.', $next));
+            }
+            if (!array_key_exists($key, $array)) {
+                throw new RuntimeException(sprintf('Key "%s" not found in "%s" (resolved from "%s")', $key, json_encode($array), $next));
+            }
+
+            return $array[$key];
+        }
 
         if ('file' === $prefix) {
             if (!is_scalar($file = $getEnv($name))) {
@@ -128,8 +149,8 @@ class EnvVarProcessor implements EnvVarProcessorInterface
                 throw new RuntimeException(sprintf('Invalid JSON in env var "%s": '.json_last_error_msg(), $name));
             }
 
-            if (!is_array($env)) {
-                throw new RuntimeException(sprintf('Invalid JSON env var "%s": array expected, %s given.', $name, gettype($env)));
+            if (null !== $env && !is_array($env)) {
+                throw new RuntimeException(sprintf('Invalid JSON env var "%s": array or null expected, %s given.', $name, gettype($env)));
             }
 
             return $env;
@@ -147,6 +168,10 @@ class EnvVarProcessor implements EnvVarProcessorInterface
 
                 return $value;
             }, $env);
+        }
+
+        if ('csv' === $prefix) {
+            return str_getcsv($env);
         }
 
         throw new RuntimeException(sprintf('Unsupported env var prefix "%s".', $prefix));