+++ src/Symfony/Component/Finder/CHANGELOG.md
@@ -1,6 +1,17 @@
 CHANGELOG
 =========
 
+4.2.0
+-----
+
+ * added $useNaturalSort option to Finder::sortByName() method
+
+4.0.0
+-----
+
+ * removed `ExceptionInterface`
+ * removed `Symfony\Component\Finder\Iterator\FilterIterator`
+
 3.4.0
 -----